Create identify-file-type.feature
First iteration of the identify file type acceptance test. This is a regression test of existing functionality. No attempt has been made yet to configure with steps.py to be able to execute the test with Behave framework. The initial focus here is on identifying the right acceptance criteria for the regression test.