added comments to clarify consent steps
Strictly speaking our scenarios do not cover all of the options a user can select when presented with the Information Release consent page. The final scenario (Preservation-User with existing IdP session logs in - Alice) was not behaving as I expected, but this was due to having chosen to have my consent information retained.  By changing the options I was able to get the test to execute succesfully.  Added comments to scenario to explain.