Create provision-user.feature
This is very similar to authenticate-dashboard-user.feature
However, given step is changed to indicate users do not yet have Archivematica accounts
And added a step towards the end for creation of the user account.  
I have already tested this script with users I created in LDAP only (dd and ee) and confirmed it works as described here.